infra-health-checker/requirements.txt
PyYAML>=6.0.1
Jinja2>=3.1.3
requests>=2.31.0
pytest>=8.0.0
pytest-cov>=4.1.0
